ai-detector/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pillow==10.4.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
